Added http_proxy based on the ENV
Fixed the gets issue by explicitly using STDIN.gets.
gets will read ARGV as a file. When a user is specified in the
command line call gets tries to load the username as a file and
expects to read in the first line of the file.

Use * instead of star(0xe29885) since it makes my putty terminal
all wonky. Will try to revert to star(0xe29885) at some point.